•  Fixed conflation of FLT_MAX with CGFLOAT_MAX (actually a double) to eliminate warnings in Xcode 9.

+++ NS(Attributed)String+Geometrics/NS(Attributed)String+Geometrics.h
@@ -128,10 +128,10 @@ extern NSInteger gNSStringGeometricsTypesetterBehavior ;
 @interface NSAttributedString (Geometrics) 
 
 // Measuring Attributed Strings
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height ;
-- (float)heightForWidth:(float)width ;
-- (float)widthForHeight:(float)height ;
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height ;
+- (CGFloat)heightForWidth:(CGFloat)width ;
+- (CGFloat)widthForHeight:(CGFloat)height ;
 
 /*!
  @brief    Returns a replica of the receiver which will fit into a given width,
@@ -139,29 +139,29 @@ extern NSInteger gNSStringGeometricsTypesetterBehavior ;
  @details  If no truncation is necessary, or if the receiver has less than two
  characters, returns the receiver (self).
  */
-- (NSAttributedString*)attributedStringTruncatedToWidth:(float)width
-                                                 height:(float)height ;
+- (NSAttributedString*)attributedStringTruncatedToWidth:(CGFloat)width
+                                                 height:(CGFloat)height ;
 
 @end
 
 @interface NSString (Geometrics)
 
 // Measuring a String With Attributes
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height
 			attributes:(NSDictionary*)attributes ;
-- (float)heightForWidth:(float)width
+- (CGFloat)heightForWidth:(CGFloat)width
 			 attributes:(NSDictionary*)attributes ;
-- (float)widthForHeight:(float)height
+- (CGFloat)widthForHeight:(CGFloat)height
 			 attributes:(NSDictionary*)attributes ;
 
 // Measuring a String with a constant Font
-- (NSSize)sizeForWidth:(float)width 
-				height:(float)height
+- (NSSize)sizeForWidth:(CGFloat)width
+				height:(CGFloat)height
 				  font:(NSFont*)font ;
-- (float)heightForWidth:(float)width
+- (CGFloat)heightForWidth:(CGFloat)width
 				   font:(NSFont*)font ;
-- (float)widthForHeight:(float)height
+- (CGFloat)widthForHeight:(CGFloat)height
 				   font:(NSFont*)font ;
 
 @end
